Make TI data available before user interaction
Before this commit the value of a ti_input would be an empty list unless
the user interacted with the ti_input widget. This is because shiny does
not call the `getValue` method on the JS binding when the component is
loaded. A conditional was added so that if the raw input value is just
an empty list (what seems to be jsonlite::fromJSON's default value) then
the initial data is returned.